# php/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-BookingHealthcare API - Swagger",
        "description": "E-BookingHealthcare API with Laravel",
        "contact": {
            "name": "Nguyen Chinh",
            "url": "https://github.com/ngchinhdev/",
            "email": "chinhnguyennn24@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/specialties": {
            "get": {
                "tags": [
                    "Specialty Routes"
                ],
                "summary": "Get all specialties",
                "operationId": "ee6f69e6cdf46cb31c5f1d57d7d82217",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/specialties/{id}": {
            "get": {
                "tags": [
                    "Specialty Routes"
                ],
                "summary": "Get one specialty",
                "operationId": "d858fad90c4bbb14808cfd1025b0f0c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/specialties/add": {
            "post": {
                "tags": [
                    "Specialty Routes"
                ],
                "summary": "Add a Specialty",
                "operationId": "e07855f375ad5d1080c5a2974e39a651",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Specialty"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "Description of the new specialty"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description of the new specialty"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/specialties/update/{id}": {
            "put": {
                "tags": [
                    "Specialty Routes"
                ],
                "summary": "Update specialty",
                "operationId": "0d484ab30a8d423b9d4358a20f4065a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Specialty"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "Description of the new specialty"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description of the new specialty"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/specialties/delete/{id}": {
            "delete": {
                "tags": [
                    "Specialty Routes"
                ],
                "summary": "Update specialty",
                "operationId": "472524189c9e91c7957f9b4932cf9d4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        }
    },
...
}